cephadm-adopt: Add python3-asyncssh to the list of required packages
The ceph-mgr-cephadm package relies on the availability of the Python
asyncssh module, but only Debian's upstream packaging gets this right:

https://packages.debian.org/unstable/admin/ceph-mgr-cephadm

Neither the Ubuntu or Fedora distro packages, nor all of the Ceph
community packages include the python3-asyncssh dependency.

Not having python3-asyncssh available on a mgr host means that the
cephadm-adopt playbook can install the ceph-mgr-cephadm package, but
then fails to enable the cephadm mgr module.

Arguably this should be fixed in upstream packaging, but until such
time it's probably prudent for the playbook to include this package in
its own dependency list.

Thus, add python3-asyncssh to the "install cephadm requirements" task.